test(field_condition): split phase-mismatch suite into its own file
Keep the unit test file under the 1k-line maintainability bar by moving the
relocation harness and Create/Update mismatch cases to a sibling package file.
