+++ internal/provider/resource_field_condition_test.go
@@ -9,7 +9,6 @@ import (
 	"io"
 	"net/http"
 	"net/http/httptest"
-	"regexp"
 	"strconv"
 	"strings"
 	"testing"
@@ -35,8 +34,7 @@ func fieldConditionBody(name string) string {
 	return fieldConditionBodyOnPhase(name, "phase_1")
 }
 
-// fieldConditionBodyOnPhase is fieldConditionBody with the owning phase spelled
-// out, so a mock can report a phase other than the one the write requested.
+// fieldConditionBodyOnPhase is fieldConditionBody with an explicit owning phase.
 func fieldConditionBodyOnPhase(name, phaseID string) string {
 	return `{"id":"fc_123","name":"` + name + `","phase":{"id":"` + phaseID + `"},` +
 		`"condition":{"expressions":[{"structure_id":"0","field_address":"1001","operation":"equals","value":"Other"}],"expressions_structure":[[0]]},` +
@@ -888,206 +886,5 @@ func TestUnit_FieldConditionResource_DestroyFallback(t *testing.T) {
 	}
 	if repoIDCalls < 2 {
 		t.Fatalf("expected repo_id to be queried again at destroy time, got %d calls", repoIDCalls)
-	}
-}
-
-// fieldConditionPhaseState drives the mocks that reproduce the API attaching a
-// condition to a phase other than the one the write asked for. Each operation
-// reports its own phase, so a read can still find the condition where it was
-// while an update relocates it.
-type fieldConditionPhaseState struct {
-	name        string
-	createPhase string
-	updatePhase string
-	readPhase   string
-	deleteFails bool
-	created     int
-	deleted     int
-}
-
-func fieldConditionPhaseHandler(st *fieldConditionPhaseState) http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		if r.Header.Get("Authorization") != "Bearer testtoken" {
-			w.WriteHeader(http.StatusUnauthorized)
-			_, _ = io.WriteString(w, `{"errors":[{"message":"unauthorized"}]}`)
-			return
-		}
-		var gr gqlReq
-		defer r.Body.Close()
-		b, _ := io.ReadAll(r.Body)
-		_ = json.Unmarshal(b, &gr)
-		w.Header().Set("Content-Type", "application/json")
-
-		if v, ok := gr.Variables["input"].(map[string]any); ok {
-			if n, ok := v["name"].(string); ok {
-				st.name = n
-			}
-		}
-
-		q := gr.Query
-		switch {
-		case strings.Contains(q, "createFieldCondition"):
-			st.created++
-			_, _ = io.WriteString(w, `{"data":{"createFieldCondition":{"fieldCondition":`+fieldConditionBodyOnPhase(st.name, st.createPhase)+`}}}`)
-		case strings.Contains(q, "updateFieldCondition"):
-			_, _ = io.WriteString(w, `{"data":{"updateFieldCondition":{"fieldCondition":`+fieldConditionBodyOnPhase(st.name, st.updatePhase)+`}}}`)
-		case strings.Contains(q, "deleteFieldCondition"):
-			st.deleted++
-			if st.deleteFails {
-				_, _ = io.WriteString(w, `{"errors":[{"message":"the phase is locked"}]}`)
-				return
-			}
-			_, _ = io.WriteString(w, `{"data":{"deleteFieldCondition":{"success":true}}}`)
-		case strings.Contains(q, "repo_id"):
-			_, _ = io.WriteString(w, `{"data":{"phase":{"repo_id":123}}}`)
-		case strings.Contains(q, "fieldCondition("):
-			_, _ = io.WriteString(w, `{"data":{"fieldCondition":`+fieldConditionBodyOnPhase(st.name, st.readPhase)+`}}`)
-		default:
-			_, _ = io.WriteString(w, `{"data":{}}`)
-		}
-	}
-}
-
-// fieldConditionPhaseConfig always asks for phase_2, the phase the relocation
-// mocks either honor or replace with the start form's phase_1.
-func fieldConditionPhaseConfig(endpoint, name string) string {
-	return `
-	provider "pipefy" {
-		endpoint = "` + endpoint + `"
-		token    = "testtoken"
-	}
-
-	resource "pipefy_field_condition" "test" {
-		phase_id = "phase_2"
-		name     = "` + name + `"
-
-		condition = {
-			all_of = [
-				{
-					field     = "1001"
-					operation = "equals"
-					value     = "Other"
-				}
-			]
-		}
-
-		actions = [
-			{
-				field     = "1002"
-				when_true = "show"
-			}
-		]
-	}
-	`
-}
-
-// TestUnit_FieldConditionResource_CreatePhaseMismatchRollsBack covers a create
-// on a phase the API does not honor: it attaches the condition to the start
-// form instead and reports that phase back. Recording that phase fails the
-// apply and taints the resource forever, so the provider deletes what it just
-// created and says why the phase cannot be managed.
-func TestUnit_FieldConditionResource_CreatePhaseMismatchRollsBack(t *testing.T) {
-	st := &fieldConditionPhaseState{createPhase: "phase_1", updatePhase: "phase_1", readPhase: "phase_1"}
-	srv := httptest.NewServer(fieldConditionPhaseHandler(st))
-	defer srv.Close()
-
-	resource.UnitTest(t, resource.TestCase{
-		TerraformVersionChecks: []tfversion.TerraformVersionCheck{
-			tfversion.SkipBelow(tfversion.Version1_8_0),
-		},
-		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
-		Steps: []resource.TestStep{
-			{
-				Config:      fieldConditionPhaseConfig(srv.URL, "Show details when type is Other"),
-				ExpectError: regexp.MustCompile(`(?s)phase_2.*phase_1.*start form`),
-			},
-		},
-	})
-
-	if st.created != 1 {
-		t.Fatalf("expected exactly 1 createFieldCondition, got %d", st.created)
-	}
-	// The one delete is the rollback: a resource left in state would add a
-	// second at destroy time, and one never created would leave none.
-	if st.deleted != 1 {
-		t.Fatalf("expected the rollback to delete the condition exactly once, got %d deletes", st.deleted)
-	}
-}
-
-// TestUnit_FieldConditionResource_CreatePhaseMismatchRollbackFails covers the
-// rollback itself failing: the condition exists on a phase Terraform is not
-// tracking, and the error has to name both problems so someone can go delete it
-// by hand.
-func TestUnit_FieldConditionResource_CreatePhaseMismatchRollbackFails(t *testing.T) {
-	st := &fieldConditionPhaseState{createPhase: "phase_1", updatePhase: "phase_1", readPhase: "phase_1", deleteFails: true}
-	srv := httptest.NewServer(fieldConditionPhaseHandler(st))
-	defer srv.Close()
-
-	resource.UnitTest(t, resource.TestCase{
-		TerraformVersionChecks: []tfversion.TerraformVersionCheck{
-			tfversion.SkipBelow(tfversion.Version1_8_0),
-		},
-		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
-		Steps: []resource.TestStep{
-			{
-				Config:      fieldConditionPhaseConfig(srv.URL, "Show details when type is Other"),
-				ExpectError: regexp.MustCompile(`(?s)orphaned.*phase_2.*rollback failed.*the phase is locked`),
-			},
-		},
-	})
-
-	if st.deleted != 1 {
-		t.Fatalf("expected exactly 1 rollback delete attempt, got %d", st.deleted)
-	}
-}
-
-// TestUnit_FieldConditionResource_UpdatePhaseMismatchFails covers an update
-// whose response reports a different phase than the one under management. The
-// condition already exists and prior state still describes it, so the update
-// fails without a rollback and a later refresh reconciles the phase.
-func TestUnit_FieldConditionResource_UpdatePhaseMismatchFails(t *testing.T) {
-	st := &fieldConditionPhaseState{createPhase: "phase_2", updatePhase: "phase_2", readPhase: "phase_2"}
-	srv := httptest.NewServer(fieldConditionPhaseHandler(st))
-	defer srv.Close()
-
-	resource.UnitTest(t, resource.TestCase{
-		TerraformVersionChecks: []tfversion.TerraformVersionCheck{
-			tfversion.SkipBelow(tfversion.Version1_8_0),
-		},
-		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
-		Steps: []resource.TestStep{
-			{
-				Config: fieldConditionPhaseConfig(srv.URL, "Show details when type is Other"),
-				ConfigStateChecks: []statecheck.StateCheck{
-					statecheck.ExpectKnownValue(
-						"pipefy_field_condition.test",
-						tfjsonpath.New("phase_id"),
-						knownvalue.StringExact("phase_2"),
-					),
-				},
-			},
-			{
-				PreConfig:   func() { st.updatePhase = "phase_1" },
-				Config:      fieldConditionPhaseConfig(srv.URL, "Renamed condition"),
-				ExpectError: regexp.MustCompile(`(?s)phase_2.*phase_1.*start form`),
-			},
-			// The failed update must not have thrown the condition away: the API
-			// cooperating again updates the same one instead of creating another.
-			{
-				PreConfig: func() { st.updatePhase = "phase_2" },
-				Config:    fieldConditionPhaseConfig(srv.URL, "Renamed condition"),
-				ConfigStateChecks: []statecheck.StateCheck{
-					statecheck.ExpectKnownValue(
-						"pipefy_field_condition.test",
-						tfjsonpath.New("name"),
-						knownvalue.StringExact("Renamed condition"),
-					),
-				},
-			},
-		},
-	})
-
-	if st.created != 1 {
-		t.Fatalf("expected the failed update to keep the condition under management, got %d creates", st.created)
 	}
 }
